started reconfig and modularisation
