+++ .env.sample
@@ -1,2 +1,3 @@
 # MongoDB-Mongoose Connection String
 DB_URI=""
+JWT_SECRET=""
