+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 package-lock.json
 .DS_Store
 .env
+.vscode
