+++ package.json
@@ -2,7 +2,7 @@
   "name": "yallambee-booking-app-backend",
   "version": "1.0.0",
   "description": "## Table of Contents",
-  "main": "index.js",
+  "main": "src/index.js",
   "type": "module",
   "directories": {
     "doc": "docs"
@@ -17,9 +17,13 @@
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.1.1",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "express-validator": "^7.2.0",
-    "mongoose": "^8.5.3"
+    "helmet": "^7.1.0",
+    "jsonwebtoken": "^9.0.2",
+    "mongoose": "^8.5.3",
+    "morgan": "^1.10.0"
   }
 }
