+++ src/controllers/homeController.js
@@ -1,0 +1,2 @@
+
+export const getHome = (req, res) => res.send("<h1>Yallambee Tiny Homes</h1>");
