+++ src/models/index.js
@@ -1,0 +1,8 @@
+import User from './User.js';
+import Property from './Property.js';
+import Booking from './Booking.js';
+
+export { User, Property, Booking };
+
+
+// Will hopefully make it easier to import models elsewhere - we can export all models from here :)
