+++ src/routes/propertyRoutes.js
@@ -1,0 +1,6 @@
+
+// Placeholder for properties listings
+const properties = ['Yallambee', 'Coming Soon']
+
+// Get properties
+app.get('/properties', (req, res) => res.send(properties))
